Create README.md
Depression is one of the most prominent problems in younger generations and one of the most difficult to detect in time. Therefore, in this GitHub we train a model capable of detecting texts with depression as a theme in posts on the social network Reddit using the Latent Dirichlet Allocation (LDA) method.